refactor: change monetary fields from float64 to int64 for consistency across booking, event, and payment services

+++ services/booking-service/cmd/migrations/000001_create_booking_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS bookings (
     user_id BIGINT NOT NULL,
     event_id BIGINT NOT NULL,
     quantity INTEGER NOT NULL,
-    total_amount DECIMAL(20, 2) NOT NULL,
+    total_amount BIGINT NOT NULL,
     status VARCHAR(50) NOT NULL DEFAULT 'PENDING',
     created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
     expires_at TIMESTAMP WITH TIME ZONE NOT NULL
