+++ services/booking-service/internal/infrastructure/clients/payment_client.go
@@ -31,7 +31,7 @@ func NewPaymentGRPCClient() (domain.PaymentProvider, error) {
 	}, nil
 }
 
-func (c *PaymentGRPCClient) CreatePayment(ctx context.Context, bookingID string, userID int64, amount float64, unitPrice float64, quantity int32) (string, error) {
+func (c *PaymentGRPCClient) CreatePayment(ctx context.Context, bookingID string, userID int64, amount int64, unitPrice int64, quantity int32) (string, error) {
 	resp, err := c.client.CreatePayment(ctx, &paymentpb.CreatePaymentRequest{
 		BookingId: bookingID,
 		UserId:    userID,
