+++ services/payment-service/cmd/migrations/000001_create_payment_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS payments (
     id VARCHAR(255) PRIMARY KEY, -- UUID string
     booking_id VARCHAR(255) NOT NULL,
     user_id BIGINT NOT NULL,
-    amount DECIMAL(20, 2) NOT NULL,
+    amount BIGINT NOT NULL,
     currency VARCHAR(10) NOT NULL DEFAULT 'idr',
     status VARCHAR(50) NOT NULL DEFAULT 'PENDING',
     stripe_id VARCHAR(255),
